Make CurrencyRateCollectionViewCell
Make CurrencyRateCollectionViewCell. This UI supports only Dark Theme since design prototype for light theme wasn't provided.

Special feature of this UI is a TimeframeOption enum. It conforms to CaseIterable protocol and has a String raw type. CurrencyRateCollectionViewCell was designed using this information, and if you will want to add some new timeframe option (for example, a "Year") to this cell, all you'll have to do is to add a case with this option to TimeframeOption enum. UI will automatically add a new button to a collection view and adjust select behavior for the new button element. You can think of TimeframeOption enum as of list of all possible options which you are free to edit and modify.
Order of cases defines the order of buttons in collection view. For example the first case in the enum will become the first button from the left, the second one will become the second button from the right, and so on, so the last enum case will be the first button from the right of collection view.