Create DataSourceSubscriber for scatter plot and AxisAdjustManager to it
Create an extension of CPTScatterPlot class with DataSourceSubscriber class. This class incapsulates work with CPTScatterPlotDataSource protocol in reactive way using Combine.
It implements Subscriber protocol, and since CPTScatterPlotDataSource protocol inherts from NSObjectProtocol we have to inherit DataSourceSubscriber from NSObject class.
Also create convenience method dataSourceSubscriber(numberOfRecords:plotSpace:animationDuration:automaticallyAdjustsAxises:) in the extension which creates and returns initialized DataSourceSubscriber class.

Depending on values which are comen from publisher on which DataSourceSubscriber is subscribed to there's complicated axis adjusting logic. So the logic was abstracted and encapsulated in AxisAdjustManager, so DataSourceSubscriber could just use it. You use it calling updateCurrentValue(_:) method, and the return value is ClosedRange, which defines the area for Y values on the graph. To initialize the manager you have to provide:
- initial value to initialize 'currentValue' property;
- origin point (default value is zero). If you want your Y-area to be above a zero level you provide your value here;
- distance, which gives an upper bound using formula upperBound = origin + distance;
- magnitude. This value defines when the manager changes its adjusting behavior. The idea is when the value change is small, we don't have to "jump" up to see the change - it's enough to move upper bound to see the whole picture. But if we have 26.3 value and the next one is 2853.7 we won't see the next value because it will be out of the graph's bounds. Even if we move the upper bound to fit all the values in the graph we will see only the last value peak and all values that are great smaller than 2853.7 (for example 26.3, 31.6, 24.9, etc.) will be not visible since scaling. So the magnitude is the measure of how a new value must be bigger comparing to the previous one to manager to choose "jump" scaling rather than upper bound scaling.
For example, if you provide 13.0 magnitude manager will change adjusting behavior when the previous value was 14.6 and a new value is bigger than 14.6*13.0 = 189.8.
Under the hood mamager decides to change adjusting behavior when newValue/currentValue >= magntitude, but the rule may change in future; you will be notified about the change.
The magnitude must be grater than zero (> 0). Otherwise, initializer throws "Index out of range" error.